requirements.txt
absl-py==0.9.0
alabaster==0.7.12
alembic==1.4.2
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
appnope==0.1.0
appscript==1.0.1
asgiref==3.3.1
asn1crypto==1.3.0
astor==0.8.1
astroid==2.3.3
astropy==4.0
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.2
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.0
bokeh==1.4.0
boto==2.49.0
Bottleneck==1.3.1
certifi==2019.3.9
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
cmake==3.17.2
colorama==0.4.3
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.6.0
conda-verify==3.4.2
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cycler==0.10.0
Cython==0.29.14
cytoolz==0.10.1
dask==2.10.0
decorator==4.4.1
defusedxml==0.6.0
distributed==2.10.0
Django==2.1.8
django-admin-rangefilter==0.3.16
django-appconf==1.0.4
django-bootstrap4==0.0.8
django-cors-headers==2.5.3
django-extensions==3.0.9
django-imagekit==4.0.2
django-mathfilters==1.0.0
django-rest-swagger==2.2.0
django-seed==0.2.2
django-suit==0.2.26
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
dlib==19.19.99
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
Faker==4.1.3
fastcache==1.1.0
filelock==3.0.12
fsspec==0.6.2
future==0.18.2
gast==0.2.2
gevent==1.4.0
git-lfs==1.6
glob2==0.7
gmpy2==2.0.8
google-pasta==0.1.8
graphviz==0.13.2
greenlet==0.4.15
grpcio==1.26.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
hypothesis==4.54.2
idna==2.8
imageio==2.6.1
imagesize==1.2.0
importlib-metadata==1.7.0
imutils==0.5.3
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.10.1
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.5
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==20.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
lief==0.9.0
llvmlite==0.31.0
locket==0.2.0
lxml==4.4.2
Mako==1.1.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==8.0.2
Mosek===-version-
mpmath==1.1.0
msgpack==0.6.1
mtcnn==0.1.0
multipledispatch==0.6.0
mysqlclient==1.4.6
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.47.0
numexpr==2.7.0
numpy==1.18.1
numpydoc==0.9.2
olefile==0.46
openapi-codec==1.3.2
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
openpyxl==3.0.3
opt-einsum==3.1.0
packaging==20.1
pandas==1.0.0
pandocfilters==1.4.2
parso==0.6.0
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.0.0
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
pprint==0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.2
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.4
pydub==0.24.0
pyflakes==2.1.1
pygame==1.9.6
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
pyodbc==4.0.27
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.3.4
pytest-arraydiff==0.3
pytest-astropy==0.7.0
pytest-astropy-header==0.1.1
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.1
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.6.0
QtPy==1.9.0
requests==2.21.0
rope==0.16.0
ruamel-yaml==0.15.87
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.5
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.13
sqlparse==0.4.1
statsmodels==0.11.0
sympy==1.5.1
tables==3.6.1
tblib==1.6.0
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-cpu-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
toolz==0.10.0
torch==1.4.0
tornado==6.0.3
tqdm==4.42.0
traitlets==4.3.3
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.3.1
wurlitzer==2.0.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlwings==0.17.0
xlwt==1.3.0
zict==1.0.0
zipp==0.6.0
